Add raw claims S3 bucket (#26)
This configuration will boot an S3 bucket:

![image](https://user-images.githubusercontent.com/2907397/79078198-744aa900-7cbb-11ea-9720-177d963cfa0f.png)

The bucket is unique to each commit, because CF errors when attempting to create an S3 bucket that already exists.

However, we should configure the Lambda to always use the same bucket for specific branches (like `master`). I'll revisit this in a future PR.

The "input" lambda, which accepts claims in JSON from an API endpoint, will write to this S3 bucket. We'll want to give that Lambda IAM permissions to write to this bucket.