Add i18n Support (English, Spanish, Chinese) (#31)
This PR adds support for i18n.

The form will now look up your current browser language, and change the form's language to match if your browser is set to one of Chinese/Spanish/English. If not, it defaults to English. The language can be updated with the language dropdown that already exists in the form. Your language configuration is stored in `localStorage`, so it will persist across page reloads.

To do this, we had to change the `form.json` format to support storing translated copy. We choose to store these translations inline so that you don't have to either a) manage multiple copies of the structure of your form or b) jump between copy and structure due to dereferencing. That looks like this:

```
{
  "pages": [
    {
      "title": "Name",
      "heading": "What is your name?",
    // ...
  ],
  // ...
},

// -->

{
  "pages": [
    {
      "title": {
        "en": "Name",
        "es": "Nombre",
        "zh": "名称"
      },
      "heading": {
        "en": "What is your name?"
        // ...
      },
    // ...
  ],
  // ...
}
```

Besides form context, there is also a good chunk of copy that is included regardless of the structure of your form, which previously wasn't stored in `form.json`. To support translations, we moved that copy to `form.json` and introduced a utility for fetching that (`getCopy("id")`).

To render translated text in app, you first fetch the language code:

```ts
import { useContext } from 'react'
import { LanguageContext } from '../contexts/language'

const { language } = useContext(LanguageContext)
```

Then, you can translate any `Copy` fields (`{ en: "...", es: "..."}`) with the `translate` utility:

```
import { translate } from '../forms'

translate(question.name, language)
```

Now, we have a few hundred instances of copy in the form already. As a first-pass, we use the Google Translate API to provide Spanish and Chinese translations. The logic for that translation is included in `scripts/translate.js`. Instructions for running that script are included in the top of that script. At a high-level, this script can be used while making iterate changes to `form.json` because it will only translate fields if no translation exists in the file. This means you can make changes to copy, introduce new questions, etc. and then just quickly re-run translate utility.

It's worth keeping in-mind that these translations are meant as a first-pass, and that we expect to find volunteers who can do a second-pass and update the copy

Resolves: https://github.com/usdigitalresponse/project-papua/issues/9