Update CSV Transformer with new claim format (#50)
Now that claims are stored as `{ metadata, questions }`, we had to update the transformer to use this format.

As part of that work, this PR adds a `yarn transform` command that will run the transformer function against your local dockerized S3.

You can run `yarn transform` to run on the previous hour's of claims. Or pipe in a JSON string as a Lambda event:

```bash
echo '{ "hour": "16" }' | yarn transform
```

This PR also fixes an off-by-one bug where the transformer was looking at the wrong hour.