Fix Lint Configuration (#54)
Apparently eslint and prettier differ in the path configuration -- for eslint passing `.` only runs on the root folder. You need to pass `**/*.{js,ts,jsx,tsx}` to run on all JS+ files.

This PR fixes CI to run on all files and lints the files that were previously missed. It also configures prettier to run in CI so that we can lint non-JS files.