Fix Selects + Back/Next Styling (#112)
This PR fixes the grommet regression for all selects, which came up in https://github.com/usdigitalresponse/project-papua/pull/111 as breaking the language select. Just a regression on Grommet's behalf which we happened to rely on.

This PR also removes the hover styling from back/next buttons (possibly another change from Grommet land?) that caused the background to turn black and hide the text.