Add Confirmation Page (#113)
This PR adds a confirmation page which renders a claim ID along with some explanatory text that shows a user that their claim is submitted.

Copy is pretty TBD, I just threw it together.

This PR also adds an environment variable we can enable in our backend to disable server-side validation (`DANGEROUS_DEMO_SKIP_VALIDATION=true`). Notably the 5MB limit still exists. We'll enable this on the `papua` site so we can demo the full flow, but any production deploys of the backend should (obv) not enable this feature.

## Show me the money

Review page:

![image](https://user-images.githubusercontent.com/2907397/80787876-87bf9600-8b3c-11ea-9d52-512a28118224.png)

Confirmation page (after submitting successfully):

![image](https://user-images.githubusercontent.com/2907397/80787851-7bd3d400-8b3c-11ea-8538-2bbdab803534.png)