Add form*.yml files to resolver function (#115)
This PR updates the resolver API function to copy over the form* files into the `dist/validation` folder. This was broken a few PRs ago when I moved the form copy logic in `compile-function` to copy into `backend/` not `amplify/` s.t. we could run the backend locally.

We just re-discovered the issue now because it's now a requirement to submit before getting to the new confirmation page.