Add support for question re-use (DAGs) via `definitions`
This commit adds support for DAGs, instead of just trees, by introducing a top-level `definitions` field which stores a list of questions and can be referenced in switches via `definition:{id}`.

We use this to allow the multi-select to be referenced from both the unemployement path + the employer relationship path without copy and pasting (which could easily lead to divergence between these two paths).

This idea is based on JSON Schema definitions syntax.