Restyle Benefits Sections (Pt. 1)
This commit lays the groundwork for a restyled benefits section. It doesn't yet include the alphabetical references, which'll be included in a future commit.

The benefits sections now render as expandable accordions with our modern branding applied.

This PR also includes a number of fixes for padding issues, which are included in this PR because of these padding-related changes were required to support the padding between sections.

There is also (alt/option)-click support for expanding/collapsing all accordions.

This also updates checkboxes to render error states with red box shadows.

This also makes the question name field optional and updates the sections field to support a name, among other metadata like background color.